moved registration file to utils
